create mt-data module and move entities to it
