+++ mt-data/src/main/java/org/zanata/magpie/model/Credential.java
