+++ mt-data/src/main/java/org/zanata/magpie/model/LocalCredential.java
