+++ mt-data/src/main/java/org/zanata/magpie/util/HashUtil.java
