+++ server/src/test/java/org/zanata/magpie/model/TranslatableHTMLNodeTest.java
@@ -6,6 +6,7 @@ import org.jsoup.nodes.Element;
 import org.jsoup.nodes.Node;
 import org.jsoup.parser.Tag;
 import org.junit.Test;
+import org.zanata.magpie.service.TranslatableHTMLNode;
 
 import java.util.HashMap;
 import java.util.Map;
